# Conflicts:
#	cmd/main.go
#	internal/handler/api.go
#	internal/handler/user.go
#	internal/service/user.go

+++ internal/domain/ticket.go
@@ -10,7 +10,7 @@ type Ticket struct {
 	FlightID  uuid.UUID `gorm:"type:uuid"                            json:"flight_id"`
 	UserID    uuid.UUID `gorm:"type:uuid"                            json:"user_id"`
 	Rank      string    `json:"rank"`
-	Price     int       `json:"price"`
+	Price     int64     `json:"price"`
 	CreatedAt time.Time `gorm:"autoCreateTime"                       json:"created_at"`
 	UpdatedAt time.Time `gorm:"autoUpdateTime"                       json:"updated_at"`
 }
